refactor(MCP): rename and update references for internal actions constants
